requirements.txt
fastapi[all]==0.39.0
kaggle==1.5.6
keras==2.3.1
numpy==1.16.4
pandas==0.22.0
Pillow==5.3.0
tensorflow==1.13.1
uvicorn==0.10.8